rename: 올바른 레이아웃을 위한 디렉토리 구조 변경
