+++ frontend/src/app/(with-header)/layout.tsx
@@ -1,0 +1,39 @@
+import Link from 'next/link';
+import localFont from 'next/font/local';
+
+import MSWComponent from '@/app/_components/MSWComponent';
+import Header from '@/app/_components/header/Header';
+import AuthContext from '@/context/AuthContext';
+
+import type { ReactNode } from 'react';
+import type { Metadata } from 'next';
+
+import '../globals.css';
+
+const Jalnan = localFont({
+  src: '../../../public/fonts/Jalnan2TTF.ttf',
+  display: 'swap',
+});
+
+export const metadata: Metadata = {
+  title: '이길로그',
+
+  // 추후 수정 필요
+  description: 'Generated by create next app',
+};
+
+export default function WithHeaderLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <html lang="ko" className={Jalnan.className}>
+      <body>
+        <MSWComponent />
+        <Header />
+        <AuthContext>{children}</AuthContext>
+      </body>
+    </html>
+  );
+}
