+++ frontend/src/app/(without-header)/login/_components/ui/LoginLogo.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function LoginLogo() {
+  return <div>LoginLogo</div>;
+}
