--- frontend/src/app/layout.tsx
@@ -1,34 +1,0 @@
-import AuthContext from '@/context/AuthContext';
-import type { Metadata } from 'next';
-
-import './globals.css';
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-import localFont from 'next/font/local';
-
-import Header from '@/app/_components/header/Header';
-import MSWComponent from './_components/MSWComponent';
-
-const Jalnan = localFont({
-  src: '../../public/fonts/Jalnan2TTF.ttf',
-  display: 'swap',
-});
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="ko" className={Jalnan.className}>
-      <body>
-        <MSWComponent />
-        <Header />
-        <AuthContext>{children}</AuthContext>
-      </body>
-    </html>
-  );
-}
