--- frontend/src/app/login/_components/ui/LoginLogo.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function LoginLogo() {
-  return (
-    <div>LoginLogo</div>
-  )
-}
